+++ cli/replicaset/custom.go
@@ -93,9 +93,9 @@ func (c *CustomInstance) Bootstrap(BootstrapCtx) error {
 	return newErrBootstrapByInstanceNotSupported(OrchestratorCustom)
 }
 
-// RolesAdd is not supported for a single instance by the Custom orchestrator.
-func (c *CustomInstance) RolesAdd(RolesChangeCtx) error {
-	return newErrRolesAddByInstanceNotSupported(OrchestratorCustom)
+// RolesChange is not supported for a single instance by the Custom orchestrator.
+func (c *CustomInstance) RolesChange(_ RolesChangeCtx, action RolesChangerAction) error {
+	return newErrRolesChangeByInstanceNotSupported(OrchestratorCustom, action)
 }
 
 // CustomApplication is an application with a custom orchestrator.
@@ -167,9 +167,10 @@ func (c *CustomApplication) Bootstrap(BootstrapCtx) error {
 	return newErrBootstrapByAppNotSupported(OrchestratorCustom)
 }
 
-// RolesAdd is not supported for an application by the Custom orchestrator.
-func (c *CustomApplication) RolesAdd(RolesChangeCtx) error {
-	return newErrRolesAddByAppNotSupported(OrchestratorCustom)
+// RolesChange is not supported for an application by the Custom orchestrator.
+func (c *CustomApplication) RolesChange(_ RolesChangeCtx,
+	action RolesChangerAction) error {
+	return newErrRolesChangeByAppNotSupported(OrchestratorCustom, action)
 }
 
 // getCustomInstanceTopology returns a topology for an instance.