+++ cli/replicaset/custom_test.go
@@ -17,7 +17,7 @@ var _ replicaset.Demoter = &replicaset.CustomInstance{}
 var _ replicaset.Expeller = &replicaset.CustomInstance{}
 var _ replicaset.VShardBootstrapper = &replicaset.CustomInstance{}
 var _ replicaset.Bootstrapper = &replicaset.CustomInstance{}
-var _ replicaset.RolesAdder = &replicaset.CustomInstance{}
+var _ replicaset.RolesChanger = &replicaset.CustomInstance{}
 
 var _ replicaset.Discoverer = &replicaset.CustomApplication{}
 var _ replicaset.Promoter = &replicaset.CustomApplication{}
@@ -25,7 +25,7 @@ var _ replicaset.Demoter = &replicaset.CustomApplication{}
 var _ replicaset.Expeller = &replicaset.CustomApplication{}
 var _ replicaset.VShardBootstrapper = &replicaset.CustomApplication{}
 var _ replicaset.Bootstrapper = &replicaset.CustomApplication{}
-var _ replicaset.RolesAdder = &replicaset.CustomApplication{}
+var _ replicaset.RolesChanger = &replicaset.CustomApplication{}
 
 func TestCustomApplication_Promote(t *testing.T) {
 	app := replicaset.NewCustomApplication(running.RunningCtx{})
@@ -63,10 +63,32 @@ func TestCustomApplication_Bootstrap(t *testing.T) {
 }
 
 func TestCustomApplication_RolesAdd(t *testing.T) {
+	cases := []struct {
+		name         string
+		changeAction replicaset.RolesChangerAction
+		errMsg       string
+	}{
+		{
+			name:         "roles add",
+			changeAction: replicaset.RolesAdder{},
+			errMsg:       `roles add is not supported for an application by "custom" orchestrator`,
+		},
+		{
+			name:         "roles remove",
+			changeAction: replicaset.RolesRemover{},
+			errMsg: `roles remove is not supported for an application by "custom"` +
+				" orchestrator",
+		},
+	}
+
 	instance := replicaset.NewCustomApplication(running.RunningCtx{})
-	err := instance.RolesAdd(replicaset.RolesChangeCtx{})
-	assert.EqualError(t, err,
-		`roles add is not supported for an application by "custom" orchestrator`)
+
+	for _, tc := range cases {
+		t.Run(tc.name, func(t *testing.T) {
+			err := instance.RolesChange(replicaset.RolesChangeCtx{}, tc.changeAction)
+			assert.EqualError(t, err, tc.errMsg)
+		})
+	}
 }
 
 func TestCustomInstance_Discovery(t *testing.T) {
@@ -452,9 +474,32 @@ func TestCustomInstance_Bootstrap(t *testing.T) {
 		`bootstrap is not supported for a single instance by "custom" orchestrator`)
 }
 
-func TestCustomInstance_RolesAdd(t *testing.T) {
+func TestCustomInstance_RolesChange(t *testing.T) {
+	cases := []struct {
+		name         string
+		changeAction replicaset.RolesChangerAction
+		errMsg       string
+	}{
+		{
+			name:         "roles add",
+			changeAction: replicaset.RolesAdder{},
+			errMsg: `roles add is not supported for a single instance by "custom"` +
+				" orchestrator",
+		},
+		{
+			name:         "roles remove",
+			changeAction: replicaset.RolesRemover{},
+			errMsg: `roles remove is not supported for a single instance by "custom"` +
+				" orchestrator",
+		},
+	}
+
 	instance := replicaset.NewCustomInstance(nil)
-	err := instance.RolesAdd(replicaset.RolesChangeCtx{})
-	assert.EqualError(t, err,
-		`roles add is not supported for a single instance by "custom" orchestrator`)
+
+	for _, tc := range cases {
+		t.Run(tc.name, func(t *testing.T) {
+			err := instance.RolesChange(replicaset.RolesChangeCtx{}, tc.changeAction)
+			assert.EqualError(t, err, tc.errMsg)
+		})
+	}
 }