+++ CHANGELOG.md
@@ -14,6 +14,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   config (3.0) or cartridge orchestrator.
 - `tt replicaset roles add`: command to add roles in the tarantool replicaset with
  cluster config (3.0) or cartridge orchestrator.
+- `tt replicaset roles remove`: command to remove roles in the tarantool replicaset with
+ cluster config (3.0) or cartridge orchestrator.
 
 ### Fixed